.github/workflows/requirements.txt
feedparser
python-telegram-bot
openai
requests
beautifulsoup4
